+++ armi/bookkeeping/db/__init__.py
@@ -64,14 +64,14 @@ import os
 from armi import runLog
 
 # re-export package components for easier import
-from armi.bookkeeping.db.database3 import Database3
+from armi.bookkeeping.db.database import Database
 from armi.bookkeeping.db.databaseInterface import DatabaseInterface
-from armi.bookkeeping.db.compareDB3 import compareDatabases
+from armi.bookkeeping.db.compareDB import compareDatabases
 from armi.bookkeeping.db.factory import databaseFactory
 
 
 __all__ = [
-    "Database3",
+    "Database",
     "DatabaseInterface",
     "compareDatabases",
     "databaseFactory",
@@ -129,7 +129,7 @@ def loadOperator(pathToDb, loadCycle, loadNode, allowMissing=False):
             "of the database."
         )
 
-    db = Database3(pathToDb, "r")
+    db = Database(pathToDb, "r")
     with db:
         # init Case here as it keeps track of execution time and assigns a reactor
         # attribute. This attribute includes the time it takes to initialize the reactor
@@ -167,7 +167,7 @@ def _getH5File(db):
     All this being said, we are probably violating this already with genAuxiliaryData,
     but we have to start somewhere.
     """
-    if isinstance(db, Database3):
+    if isinstance(db, Database):
         return db.h5db
     else:
         raise TypeError("Unsupported Database type ({})!".format(type(db)))
