+++ armi/reactor/parameters/parameterDefinitions.py
@@ -129,7 +129,7 @@ class Serializer:
     operations to be performed on the parameter values as they are stored to the
     database or read back in.
 
-    The ``Database3`` already knows how to handle certain cases where the data are not
+    The ``Database`` already knows how to handle certain cases where the data are not
     straightforward to get into a numpy array, such as when:
 
       - There are ``None``\ s.
@@ -169,8 +169,8 @@ class Serializer:
 
     See Also
     --------
-    armi.bookkeeping.db.database3.packSpecialData
-    armi.bookkeeping.db.database3.unpackSpecialData
+    armi.bookkeeping.db.database.packSpecialData
+    armi.bookkeeping.db.database.unpackSpecialData
     armi.reactor.flags.FlagSerializer
     """
